+++ tests/in/glsl/constant-array-size.vert
@@ -12,3 +12,5 @@ vec4 function() {
     }
     return sum;
 }
+
+void main() {}
